some progress, and fucking _EnvLegendreCut := 1 .. infinity

+++ treekernels.h
@@ -35,7 +35,7 @@ namespace ispc { /* namespace */
 extern "C" {
 #endif // __cplusplus
     extern void minmax(const int32_t N, const double * x, const double * y, const double * z, double &xmin, double &xmax, double &ymin, double &ymax, double &zmin, double &zmax);
-    extern void morton(const int32_t N, const double * x, const double * y, const double * z, const double xmin, const double ymin, const double zmin, const double ext, int32_t * index, const int32_t ntasks);
+    extern void morton(const int32_t N, const double * x, const double * y, const double * z, const double xmin, const double ymin, const double zmin, const double ext, int64_t * index, const int32_t ntasks);
     extern void node_setup(const double * xs, const double * ys, const double * zs, const double * qs, const int32_t n, double &Q, double &xcom, double &ycom, double &zcom, double &radius, double &weight);
 #if defined(__cplusplus) && (! defined(__ISPC_NO_EXTERN_C) || !__ISPC_NO_EXTERN_C )
 } /* end extern C */
